+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+[unreleased]
+- Fixed: Fix some timeout issues with the CMM workbench when using the Makera protocol and improve unexpected states detection
+
 [2.2.0-RC1]
 - Enhancement: Read tool definitions from post-processor outputs and use them in the G-code viewer
 - Enhancement: Add multi-select to the remote file browser